requirements.txt.txt
requests
cryptography
python-telegram-bot
